fix: change generate location, update composer
