+++ php/example/composer.json
@@ -1,22 +1,22 @@
 {
-    "name": "extism/example",
-    "description": "Example running the Extism PHP Host SDK",
-    "license": "BSD-3-Clause",
-    "type": "project",
-    "authors": [
-        {
-            "name": "The Extism Authors",
-            "email": "oss@extism.org"
-        }
-    ],
-    "require": {
-        "extism/extism": "*"
-    },
-    "repositories": [
-        {
-            "type": "path",
-            "url": "../../"
-        }
-    ],
-    "minimum-stability": "dev"
+  "name": "extism/example",
+  "description": "Example running the Extism PHP Host SDK",
+  "license": "BSD-3-Clause",
+  "type": "project",
+  "authors": [
+    {
+      "name": "The Extism Authors",
+      "email": "oss@extism.org"
+    }
+  ],
+  "require": {
+    "extism/extism": "*"
+  },
+  "repositories": [
+    {
+      "type": "path",
+      "url": "../../"
+    }
+  ],
+  "minimum-stability": "dev"
 }
