+++ php/src/Plugin.php
@@ -2,34 +2,8 @@
 declare(strict_types=1);
 namespace Extism;
 
-require_once "vendor/autoload.php";
 require_once "ExtismLib.php";
 
-function generate_extism_lib() {
-    return (new FFIMe\FFIMe("libextism.".soext()))
-        ->include("extism.h")
-        ->showWarnings(false)
-        ->codeGen('ExtismLib', __DIR__.'/ExtismLib.php');
-}
-
-function soext() {
-    $platform = php_uname("s");
-    switch ($platform) {
-        case "Darwin":
-            return "dylib";
-        case "Linux": 
-            return "so";
-        case "Windows":
-            return "dll";
-        default:
-            throw new Exeception("Extism: unsupported platform ".$platform);
-    }
-}
-
-if (!file_exists(__DIR__."/ExtismLib.php")) {
-    generate_extism_lib();
-}
-    
 class Plugin
 {
     private $lib;
